# doc/api.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: WASA Photo REST API
  description: |
    This is the API specification for the WASA Photo REST API.

    It's part of the WASA Photo project for the Web and Software Architecture course at Sapienza University of Rome.

    The project description and details are availabe at [this link](http://gamificationlab.uniroma1.it/notes/Project.pdf)
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      description: |
        This security scheme is used to authenticate a user.
        The user must be signed in to perform the request.
        The user must pass the bearer token in the Authorization header.
      type: http
      scheme: bearer
  schemas:
    username:
      description: Username of a user
      type: string
      minLength: 4
      maxLength: 20
      pattern: '^[a-zA-Z0-9_-]+$'
      readOnly: true
      example: mariorossi99
    resourceId:
      description: The Universally Unique Identifier (follows the standard)
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
      readOnly: true
    date:
      description: representation of Date and time
      type: string
      format: date-time
      example: 2023-01-01T00:00:00Z
      readOnly: true
    caption:
      description: A caption for a post or a comment
      type: string
      minLength: 1
      maxLength: 800
      pattern: '.*'
      example: "This is a caption"
      readOnly: false
    image:
      description: An uri to an image
      type: string
      format: uri
      example: "localhost:8080/users/mariorossi99/posts/1/image"
      readOnly: false
    counter:
      description: A simple counter,
                    keeps track of the number of followers, likes, etc.
      type: integer
      example: 8
      readOnly: true
    userCollection:
      description: A list of usersnames, 
                   it is used to keep track of the users that liked a post or a comment
      type: array
      minLength: 0
      maxLength: 69420
      items:
        anyOf:
          - $ref: '#/components/schemas/User/properties/username'
          - $ref: '#/components/schemas/User/properties/userId'
          - $ref: '#/components/schemas/User/properties/profileImage'
      readOnly: false
    User:
      title: User
      type: object
      description: A user of the WASA Photo platform
      properties:
        userId:
          $ref: '#/components/schemas/resourceId'
        username:
          $ref: '#/components/schemas/username'
        signUpDate:
          $ref: '#/components/schemas/date'
        lastSeenDate:
          $ref: '#/components/schemas/date'
        bio:
          $ref: '#/components/schemas/caption'
        profileImage:
          $ref: '#/components/schemas/image'
        followers:
          $ref: '#/components/schemas/counter'
        following:
          $ref: '#/components/schemas/counter'
      required:
        - username
        - signUpDate
...
    postStream:
      description: An array of posts, 
                   can either be the list of posts of a user or the list of posts of the users followed by a user 
      type: array
      minLength: 0
      maxLength: 69420
      items:
        $ref: '#/components/schemas/UserPost/properties/postId'
...
    Error:
      description: An error message
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]$'
      required:
        - message
    Success:
      description: A successful request
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]$'
        body:
          type: object
      required:
        - message
  parameters:
    userId:
      name: userId
      in: path
      description: The userId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
...
  responses:
...
    NotFound: #for 404
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    BadRequest: #for 400
      description: The request was not valid, 
                    the request body is missing or malformed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError: #for 500
      description: An error occurred on the server side
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict: #for 409
      description: The request could not be completed due to a conflict with the current state of the target resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    
    
    
    #Success responses
...
    Ok: #for 200
      description: The request has succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
...
paths:
...
  /users:
    description: This endpoints handles collection of users.
    
    get:
      tags: ["user"]
      operationId: searchUser
      summary: Search for a user, based on the username
      description: |
        This endpoint is used to search for a user, based on the username.
        The username is passed as a query parameter.
        The response will retun a list of usernames similar to the one passed as a query parameter.
      parameters:
        - name: username
          in: query
          description: The username to search for
          required: true
          schema:
            $ref: '#/components/schemas/username'
      responses:
        "200":
          description: The list of usernames similar to the one passed as a query parameter
          content:
            application/json:
              schema: 
                $ref: '#/components/schemas/userCollection'
        "400": #the request body is missing or malformed
          $ref: '#/components/responses/BadRequest'
  /users/{userId}:
    description: This endpoints handles a single user.
                  it can be used to get the details of a user, delete it, or update it.
    parameters:
      - $ref: '#/components/parameters/userId'

    get: #get user profile
      tags: ["user"]
      operationId: getUserProfile
      summary: Get the profile of a user, it will return user details and the list of posts of the user
      description: |
        This endpoint is used to get the profile of a user.
        The userId is passed as a path parameter.
        The response will retun:
          - the details of the user
          - the list of posts of the user.
      responses:
        "200":
          description: User is found and profile is returned in the response body
          content:
            application/json:
              schema:
                type: object
                properties:
                  userInfo:
                    $ref: '#/components/schemas/User'
                  userStream:
                    $ref: '#/components/schemas/postStream'
        "404": # user not found
          {$ref: '#/components/responses/NotFound'}
        "500":
          {$ref: '#/components/responses/InternalServerError'}

    put: #update user information
      tags: ["user"]
      operationId: setMyUserName #actually updateUserProfile since it will update any profile info
      summary: Update the profile of a user
      description: Updates the given profile
      requestBody:
        description: A user object
        content: 
          application/json:
            schema: 
              $ref: '#/components/schemas/User'
      responses:
        "202":
          description: User fields updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "400": #the request body is missing or malformed
          $ref: '#/components/responses/BadRequest'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
        "500":  
          $ref: '#/components/responses/InternalServerError'
        "409":
          $ref: '#/components/responses/Conflict'
           
    delete: #delete
      tags: ["user"]
      operationId: deleteUser
      summary: Delete a user
      description: Deletes the given user
      responses:
        "200": #user deleted
          $ref: '#/components/responses/Ok'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
...
security:
  - bearerAuth: [] 